Add Queue & Playlist features, Rename Dialog, Style, Animations, Fixes

+++ src/ui/views/Playlist/Playlist.tsx
@@ -10,7 +10,7 @@ import Lottie, { LottieRefCurrentProps } from 'lottie-react';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { removePlaylist, selectPlaylists, updatePlaylist } from '../../../state/slices/playlistSlice';
 import { PlaylistData, TrackData } from '../../../typings/playlist';
-import { setQueue } from '../../../state/slices/playerSlice';
+import { selectQueue, setQueue, updateQueue } from '../../../state/slices/playerSlice';
 import useContextMenu from '../../hooks/useContextMenu';
 import AppRoutes from '../../routes';
 
@@ -29,8 +29,13 @@ import deleteIcon from '../../../../assets/ui-icons/trash-2.svg';
 import editIcon from '../../../../assets/ui-icons/edit-3.svg';
 import addTopIcon from '../../../../assets/ui-icons/add-top.svg';
 import addBottomIcon from '../../../../assets/ui-icons/add-bottom.svg';
+import imageIcon from '../../../../assets/ui-icons/image-regular.svg';
 
 import './Playlist.scss';
+import RenameDialog from '../../components/RenameDialog/RenameDialog';
+import Dialog from '../../components/Dialog/Dialog';
+import HamburgerMenu from '../../components/HamburgerMenu/HamburgerMenu';
+import HamburgerMenuItem from '../../components/HamburgerMenu/HamburgerMenuItem/HamburgerMenuItem';
 
 const Playlist: FC = memo(function Playlist() {
 	const { id } = useParams();
@@ -39,6 +44,7 @@ const Playlist: FC = memo(function Playlist() {
 	const dispatch = useAppDispatch();
 
 	const playlists = useAppSelector(selectPlaylists);
+	const queue = useAppSelector(selectQueue);
 	const [playlist, setPlaylist] = useState(playlists.find((x) => x.id === id));
 
 	const [tracks, setTracks] = useState<TrackData[]>([]);
@@ -49,7 +55,12 @@ const Playlist: FC = memo(function Playlist() {
 		})
 	);
 
+	const [renameVisibility, setRenameVisibility] = useState(false);
+	const [hamburgerVisibility, setHamburgerVisibility] = useState(false);
+	const [isDialogVisible, setDialogVisibility] = useState(false);
+
 	const [visibility, setVisibility, position, setPosition] = useContextMenu();
+	const [usingContextMenuId, setUsingContextMenuId] = useState('');
 
 	const lottieShowMenuRef = useRef<LottieRefCurrentProps>(null);
 	const lottieAddTracksRef = useRef<LottieRefCurrentProps>(null);
@@ -77,11 +88,13 @@ const Playlist: FC = memo(function Playlist() {
 		}
 	};
 
-	const deletePlaylist = () => {
+	const deletePlaylist = (confirm: boolean) => {
 		if (!playlist?.id) return;
 
-		dispatch(removePlaylist(playlist.id));
-		navigate(AppRoutes.Library);
+		if (confirm) {
+			dispatch(removePlaylist(playlist.id));
+			navigate(AppRoutes.Library);
+		}
 	};
 
 	const handleAddTracks = async () => {
@@ -133,6 +146,51 @@ const Playlist: FC = memo(function Playlist() {
 		e?.current?.play();
 	};
 
+	const handleTrackRemove = (id: string) => {
+		if (!playlist) return;
+
+		const updateData: PlaylistData = {
+			id: playlist.id,
+			name: playlist.name,
+			tracks: [...playlist.tracks],
+			pinned: playlist.pinned
+		};
+
+		const index = updateData.tracks.findIndex((x) => x.id === id);
+		updateData.tracks.splice(index, 1);
+
+		dispatch(updatePlaylist(updateData));
+		setTracks(updateData.tracks);
+	};
+
+	const handlePlayNext = (id: string) => {
+		if (!playlist) return;
+
+		const updateData: TrackData[] = [...queue];
+
+		const index = playlist.tracks.findIndex((x) => x.id === id);
+		const track = playlist.tracks[index];
+		updateData.splice(1, 0, track);
+
+		dispatch(updateQueue(updateData));
+	};
+
+	const handlePlayLast = (id: string) => {
+		if (!playlist) return;
+
+		const updateData: TrackData[] = [...queue];
+
+		const index = playlist.tracks.findIndex((x) => x.id === id);
+		const track = playlist.tracks[index];
+		updateData.push(track);
+
+		dispatch(updateQueue(updateData));
+	};
+  
+  const handleRename = (data: string) => {
+		if (playlist) dispatch(updatePlaylist({ ...playlist, name: data } as PlaylistData));
+	};
+
 	useEffect(() => {
 		const playlistFound = playlists.find((x) => x.id === id);
 		if (playlistFound) {
@@ -141,12 +199,31 @@ const Playlist: FC = memo(function Playlist() {
 		}
 	}, [playlists, id]);
 
+	useEffect(() => {
+		if (lottieShowMenuRef.current) {
+			lottieShowMenuRef.current.setSpeed(2);
+			if (!hamburgerVisibility) {
+				stopAnimation(lottieShowMenuRef);
+			} else {
+				startAnimation(lottieShowMenuRef);
+			}
+		}
+	}, [hamburgerVisibility]);
+
 	return (
 		<div key={id} id="playlist-container">
+			<Dialog heading="Delete?" description="You are about to delete this playlist. This action cannot be undone!" onClose={() => setDialogVisibility(false)} isOpen={isDialogVisible} type="danger" confirmText="Delete" rejectText="Keep" confirmCallback={deletePlaylist} />
 			<div id="playlist-heading">
+				<RenameDialog value={playlist?.name} nameCB={handleRename} visible={renameVisibility} onClose={() => setRenameVisibility(false)} />
 				<img src={defaultAlbumCover} alt="" />
-				<div id="playlist-heading-text">{playlist?.name}</div>
+				<div id="playlist-heading-text" onClick={() => setRenameVisibility(true)}>
+					{playlist?.name}
+				</div>
 				<div id="playlist-controls">
+					<HamburgerMenu onClose={() => setHamburgerVisibility(false)} visible={hamburgerVisibility} positionX={25} positionY={100}>
+						<HamburgerMenuItem title="Edit" icon={editIcon} />
+						<HamburgerMenuItem title="Choose image" icon={imageIcon} />
+					</HamburgerMenu>
 					<ToolTip text="Play Playlist">
 						<div id="playlist-play-btn" className="playlist-heading-btn btn-hover-animation" onClick={playPlaylist}>
 							<img id="play-icon" src={playIcon} alt="" draggable="false" />
@@ -158,13 +235,15 @@ const Playlist: FC = memo(function Playlist() {
 						</div>
 					</ToolTip>
 					<ToolTip text="Delete Playlist">
-						<div id="playlist-delete-btn" className="playlist-heading-btn btn-hover-animation" onClick={deletePlaylist} onMouseEnter={() => startAnimation(lottieDeletePlaylistRef)} onMouseLeave={() => stopAnimation(lottieDeletePlaylistRef)}>
+						<div id="playlist-delete-btn" className="playlist-heading-btn btn-hover-animation" onClick={() => setDialogVisibility(true)} onMouseEnter={() => startAnimation(lottieDeletePlaylistRef)} onMouseLeave={() => stopAnimation(lottieDeletePlaylistRef)}>
 							<Lottie id="delete-icon" animationData={trashIcon} lottieRef={lottieDeletePlaylistRef} loop={false} autoplay={false} />
 						</div>
 					</ToolTip>
-					<div id="playlist-menu-btn" className="playlist-heading-btn btn-hover-animation">
-						<Lottie id="menu-icon" animationData={menuIcon} lottieRef={lottieShowMenuRef} loop={false} autoplay={false} />
-					</div>
+					<ToolTip text="Options">
+						<div id="playlist-menu-btn" className="playlist-heading-btn btn-hover-animation" onClick={() => setHamburgerVisibility(true)}>
+							<Lottie id="menu-icon" animationData={menuIcon} lottieRef={lottieShowMenuRef} loop={false} autoplay={false} />
+						</div>
+					</ToolTip>
 				</div>
 			</div>
 			<div id="divider" />
@@ -180,6 +259,7 @@ const Playlist: FC = memo(function Playlist() {
 										track={track}
 										onContextMenu={(e) => {
 											e.preventDefault();
+											setUsingContextMenuId(track.id);
 											setVisibility(true);
 											setPosition({
 												x: e.pageX,
@@ -193,14 +273,14 @@ const Playlist: FC = memo(function Playlist() {
 						</SortableContext>
 					</DndContext>
 				)}
+				{visibility && (
+					<ContextMenu y={position.y} x={position.x}>
+						<ContextMenuItem header="Play Next" staticIcon={addTopIcon} type="default" onClick={() => handlePlayNext(usingContextMenuId)} />
+						<ContextMenuItem header="Play Last" staticIcon={addBottomIcon} type="default" onClick={() => handlePlayLast(usingContextMenuId)} />
+						<ContextMenuItem header="Delete from Playlist" staticIcon={deleteIcon} type="danger" onClick={() => handleTrackRemove(usingContextMenuId)} />
+					</ContextMenu>
+				)}
 			</div>
-			{visibility && (
-				<ContextMenu y={position.y} x={position.x}>
-					<ContextMenuItem header="Play Next" staticIcon={addTopIcon} type="default" />
-					<ContextMenuItem header="Play Later" staticIcon={addBottomIcon} type="default" />
-					<ContextMenuItem header="Delete from Playlist" staticIcon={deleteIcon} type="danger" />
-				</ContextMenu>
-			)}
 		</div>
 	);
 });
